Version 0.7.0 - replaced unsigned long with uint64_t and long with int64_t in all cpp files - per Google C++ Style Guide

+++ src/cpp/GanitaNode.cpp
@@ -8,7 +8,7 @@ GanitaNode::GanitaNode(void)
   value = 0;
 }
 
-GanitaNode::GanitaNode(unsigned long ii, unsigned val, string edgeval)
+GanitaNode::GanitaNode(uint64_t ii, unsigned val, string edgeval)
 {
   delim = EDGE_STR_DELIM;
   id = ii;
@@ -23,12 +23,12 @@ GanitaNode::GanitaNode(unsigned long ii, unsigned val, string edgeval)
   }
 }
 
-unsigned long GanitaNode::returnId(void)
+uint64_t GanitaNode::returnId(void)
 {
   return(id);
 }
 
-unsigned long GanitaNode::returnValue(void)
+uint64_t GanitaNode::returnValue(void)
 {
   return(value);
 }
@@ -38,47 +38,47 @@ string GanitaNode::returnProperty(void)
   return(*property);
 }
 
-unsigned long GanitaNode::initProperty(string prop)
+uint64_t GanitaNode::initProperty(string prop)
 {
   property = new string(prop);
   return(property->size());
 }
 
-unsigned long GanitaNode::setId(unsigned long ii)
+uint64_t GanitaNode::setId(uint64_t ii)
 {
   id = ii;
 
   return(id);
 }
 
-unsigned long GanitaNode::setValue(unsigned long val)
+uint64_t GanitaNode::setValue(uint64_t val)
 {
   value = val;
 
   return(value);
 }
 
-unsigned long GanitaNode::addEdge(unsigned long val)
+uint64_t GanitaNode::addEdge(uint64_t val)
 {
   GanitaEdge tmpEdge(edge.size() + 1, val);
 
   edge.push_back(tmpEdge);
 
-  return((unsigned long)edge.size());
+  return((uint64_t)edge.size());
 }
 
-unsigned long GanitaNode::addEdge(unsigned long ii, unsigned long val)
+uint64_t GanitaNode::addEdge(uint64_t ii, uint64_t val)
 {
   GanitaEdge tmpEdge(ii, val);
 
   edge.push_back(tmpEdge);
 
-  return((unsigned long)edge.size());
+  return((uint64_t)edge.size());
 }
 
 // Similar to previous function but estimates 
 // temporary memory for token. 
-unsigned long GanitaNode::setEdge(string edgeval){
+uint64_t GanitaNode::setEdge(string edgeval){
   delim = EDGE_STR_DELIM;
   int jj; jj = 0;
   string token;
@@ -89,13 +89,13 @@ unsigned long GanitaNode::setEdge(string edgeval){
     jj++;
   }
 
-  return((unsigned long)edge.size());
+  return((uint64_t)edge.size());
 }
 
-unsigned long GanitaNode::edgestrToMax(string edgeval)
+uint64_t GanitaNode::edgestrToMax(string edgeval)
 {
-  long double tmp;
-  tmp = (long double) edgeval.size() + 1;
+  double tmp;
+  tmp = (double) edgeval.size() + 1;
 
   if((size_t) tmp != edgeval.size()){
     fprintf(stderr, "Unable to allocate edges.\n");
@@ -105,35 +105,35 @@ unsigned long GanitaNode::edgestrToMax(string edgeval)
   // Add 1 to round up and add another 1 for the comma delimiters in edgeval.
   tmp = ceil(tmp/(log10(tmp) + 1)) + 2;
   
-  return((unsigned long) tmp);
+  return((uint64_t) tmp);
 }
 
-unsigned long GanitaNode::dumpEdgeValues(void)
+uint64_t GanitaNode::dumpEdgeValues(void)
 {
   if(edge.size() <= 0) return(0);
-  unsigned long ii;
+  uint64_t ii;
   cout<<edge[0].returnValue();
-  for(ii=1; ii<(unsigned long)edge.size(); ii++){
+  for(ii=1; ii<(uint64_t)edge.size(); ii++){
     cout<<","<<edge[ii].returnValue();
   }
 
-  return((unsigned long) edge.size());
+  return((uint64_t) edge.size());
 }
 
-unsigned long GanitaNode::dumpEdgeAll(void)
+uint64_t GanitaNode::dumpEdgeAll(void)
 {
   if(edge.size() <= 0) return(0);
-  unsigned long ii;
+  uint64_t ii;
   cout<<edge[0].returnId()<<","<<edge[0].returnValue();
-  for(ii=1; ii<(unsigned long)edge.size(); ii++){
+  for(ii=1; ii<(uint64_t)edge.size(); ii++){
     cout<<"|"<<edge[ii].returnId()<<","<<edge[ii].returnValue();
   }
   cout<<endl;
 
-  return((unsigned long) edge.size());
+  return((uint64_t) edge.size());
 }
 
-int GanitaNode::returnEdge(unsigned long ee, GanitaEdge& ed)
+int GanitaNode::returnEdge(uint64_t ee, GanitaEdge& ed)
 {
   if(edge.size() > 0){
     ed = edge[ee % edge.size()];
@@ -142,7 +142,7 @@ int GanitaNode::returnEdge(unsigned long ee, GanitaEdge& ed)
   return(0);
 }
 
-int GanitaNode::setEdgeId(unsigned long ee, unsigned long ii)
+int GanitaNode::setEdgeId(uint64_t ee, uint64_t ii)
 {
   if(edge.size() > 0){
     edge[ee % edge.size()].setId(ii);
@@ -151,7 +151,7 @@ int GanitaNode::setEdgeId(unsigned long ee, unsigned long ii)
   return(0);
 }
 
-int GanitaNode::setEdgeValue(unsigned long ee, unsigned long val)
+int GanitaNode::setEdgeValue(uint64_t ee, uint64_t val)
 {
   if(edge.size() > 0){
     edge[ee % edge.size()].setValue(val);
@@ -160,7 +160,7 @@ int GanitaNode::setEdgeValue(unsigned long ee, unsigned long val)
   return(0);
 }
 
-unsigned long GanitaNode::returnNumEdges(void)
+uint64_t GanitaNode::returnNumEdges(void)
 {
   return(edge.size());
 }
